requirements.txt
torch==2.4.1
transformers==4.35.2
gradio==5.0.2
python-dotenv==1.0.1